Can get POSIX time, fixed WiFly_Send
Updated WiFly_Send to allow better handling of commands with no
response. A NULL will ignore any response. However, this causes the
buffer to be cleared unecessarily. Will handle this later.

Finished WiFly_Time, properly gets the time from the WiFly chip, which
gets it from an internet time server. Need to add a rotation for the
time servers it calls to make it more robust using the WiFly "lookup"
command to get DNS for the NIST time servers.